chore: 프로젝트 구조 변경
Spring 프로젝트 상위 디렉터리로 이동
.gitignore 추가
테스트 파일 삭제
