--- guess-hangeul/src/main/java/com/estsoft/guesshangeul/DTO/entity/Board.java
@@ -1,34 +1,0 @@
-package com.estsoft.guesshangeul.DTO.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Entity
-@NoArgsConstructor
-@Table
-public class Board {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "board_id", unique = true, nullable = false)
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    @ColumnDefault("1")
-    private int type; //게시판 속성
-
-    @Column(nullable = false)
-    private Long userId;
-
-    @Column(name = "created_at")
-    private LocalDateTime createTime;
-}
