--- guess-hangeul/src/main/java/com/estsoft/guesshangeul/DTO/entity/Comment.java
@@ -1,39 +1,0 @@
-package com.estsoft.guesshangeul.DTO.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-@Table(name = "comment")
-public class Comment {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id",unique = true ,nullable = false)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private Users user;
-
-    @ManyToOne
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Column(nullable = false)
-    @ColumnDefault("1")
-    private int type; //게시판 속성
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-}
