--- guess-hangeul/src/main/java/com/estsoft/guesshangeul/DTO/entity/Post.java
@@ -1,51 +1,0 @@
-package com.estsoft.guesshangeul.DTO.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Entity
-@NoArgsConstructor
-@Table
-public class Post {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_id", unique = true, nullable = false)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private Users user;
-
-    @ManyToOne
-    @JoinColumn(name = "board_id")
-    private Board board;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Column
-    private String answer;
-
-    @Column(name = "is_hidden", nullable = false)
-    @ColumnDefault("false")
-    private boolean isHidden;
-
-    @Column
-    private Long view;
-
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-}
