+++ src/main/java/com/estsoft/guesshangeul/GuessHangeulApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class GuessHangeulApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(GuessHangeulApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(GuessHangeulApplication.class, args);
+	}
 
 }
