+++ src/main/java/com/estsoft/guesshangeul/entity/Board.java
@@ -1,0 +1,40 @@
+package com.estsoft.guesshangeul.entity;
+
+import java.time.LocalDateTime;
+
+import org.hibernate.annotations.ColumnDefault;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+@NoArgsConstructor
+@Table
+public class Board {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "board_id", unique = true, nullable = false)
+	private Long id;
+
+	@Column(nullable = false)
+	private String title;
+
+	@Column(nullable = false)
+	@ColumnDefault("1")
+	private int type; //게시판 속성
+
+	@Column(nullable = false)
+	private Long userId;
+
+	@Column(name = "created_at")
+	private LocalDateTime createTime;
+}
