+++ src/main/java/com/estsoft/guesshangeul/entity/Comment.java
@@ -1,0 +1,47 @@
+package com.estsoft.guesshangeul.entity;
+
+import java.time.LocalDateTime;
+
+import org.hibernate.annotations.ColumnDefault;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name = "comment")
+public class Comment {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "comment_id", unique = true, nullable = false)
+	private Long id;
+
+	@ManyToOne
+	@JoinColumn(name = "user_id")
+	private Users user;
+
+	@ManyToOne
+	@JoinColumn(name = "post_id")
+	private Post post;
+
+	@Column(nullable = false)
+	private String content;
+
+	@Column(nullable = false)
+	@ColumnDefault("1")
+	private int type; //게시판 속성
+
+	@Column(nullable = false)
+	private LocalDateTime createdAt;
+}
